celery.py and tasks.py created and views.py, urls.py, serializers.py, authentication.py and README.md are updated.

+++ Notes_API/api/views.py
@@ -7,7 +7,6 @@ from rest_framework.permissions import IsAuthenticated
 from django.conf import settings
 from .serializers import RegisterationSerializer, LoginSerializer, NoteSerializer, CategorySerializer  
 from .models import Note, Category
-from django.shortcuts import get_object_or_404
 from .authentication import JWTAuthentication
 from rest_framework.permissions import AllowAny
 
@@ -67,58 +66,46 @@ class CreateNoteView(APIView):
         all_notes = Note.objects.filter(owner=request.user)
         serializer = NoteSerializer(all_notes, many=True, context={'request': request}) 
 
-        if all_notes:
-            return Response({"status": "success", "data": serializer.data}, status=status.HTTP_200_OK)
+        return Response({"status": "success", "data": serializer.data}, status=status.HTTP_200_OK)
     
 
 class NoteDetailView(APIView):
     authentication_classes = [JWTAuthentication]
     permission_classes = [IsAuthenticated]
 
-    def get(self, request, note_id):
-       
-        note = get_object_or_404(Note, id=note_id, owner=request.user)
-        serializer = NoteSerializer(note, context={'request': request})
-
-        return Response(serializer.data, status=status.HTTP_200_OK)
-    
-    def put(self, request, note_id):
-        try:
-            note = Note.objects.get(id=note_id, owner=request.user)
-            serializer = NoteSerializer(note, data=request.data, context={'request': request})
-            if serializer.is_valid():
-                serializer.save()
-                return Response({'msg': 'Note updated successfully', 'data': serializer.data}, status=status.HTTP_200_OK)
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-        except Note.DoesNotExist:
-            return Response({"status": "error", "message": "Note not found"}, status=status.HTTP_404_NOT_FOUND)
-
-
-    def delete(self, request, note_id):
-        try:
-            note = Note.objects.get(id=note_id, owner=request.user)
-            note.delete()
-
-            return Response({'msg': 'Note deleted successfully'}, status=status.HTTP_204_NO_CONTENT)        
-        except Note.DoesNotExist:
-            return Response({"status": "error", "message": "Note not found"}, status=status.HTTP_404_NOT_FOUND)
-        
-
-class NoteTitleSearchView(APIView):
-    authentication_classes = [JWTAuthentication]
-    permission_classes = [IsAuthenticated]
-    serializer_class = NoteSerializer
-
     def get(self, request, title):
+       
         try:
             note = Note.objects.filter(title__icontains=title, owner=request.user)
             serializer = NoteSerializer(note, many=True, context={'request': request})
 
             return Response(serializer.data, status=status.HTTP_200_OK)
-        
         except Note.DoesNotExist:
-            return Response({"error": "Note not found"}, status=status.HTTP_404_NOT_FOUND)
+            return Response({"status": "error", "message": "Note with the provided title not found"}, status=status.HTTP_404_NOT_FOUND)
+
+    
+    def put(self, request, title):
+        try:
+            note = Note.objects.get(title__icontains=title, owner=request.user)
+            serializer = NoteSerializer(note, data=request.data, context={'request': request})
+            if serializer.is_valid():
+                serializer.save()
+                return Response({"status": "success", 'data': serializer.data}, status=status.HTTP_200_OK)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        except Note.DoesNotExist:
+            return Response({"status": "error", "message": "Note with the provided title not found"}, status=status.HTTP_404_NOT_FOUND)
+
+
+    def delete(self, request, title):
+        try:
+            note = Note.objects.filter(title__icontains=title, owner=request.user)
+            if note.exists():
+                note.delete()
+                return Response({"status": "success", 'msg': 'Note deleted successfully'}, status=status.HTTP_204_NO_CONTENT)
+        except Note.DoesNotExist:
+            return Response({"status": "error", "message": "Note with the provided title not found"}, status=status.HTTP_404_NOT_FOUND)
         
+
 
 class NoteCategorySearchView(APIView):
     authentication_classes = [JWTAuthentication]
@@ -133,7 +120,18 @@ class NoteCategorySearchView(APIView):
             return Response({"status": "success", "data": serializer.data}, status=status.HTTP_200_OK)
         
         except Note.DoesNotExist:
-            return Response({"status": "error", "message": "Note not found"}, status=status.HTTP_404_NOT_FOUND)
+            return Response({"status": "error", "message": f"Notes with {category} category not found"}, status=status.HTTP_404_NOT_FOUND)
+        
+
+    def delete(self, request, category):
+        try:
+            note = Note.objects.filter(category__name__icontains=category, category__owner=request.user)
+            if note.exists():
+                note.delete()
+                return Response({"status": "success", 'msg': 'Note deleted successfully'}, status=status.HTTP_204_NO_CONTENT)
+
+        except Note.DoesNotExist:
+            return Response({"status": "error", "message": f"Notes with {category} category not found"}, status=status.HTTP_404_NOT_FOUND)
         
 
 class CategoryView(APIView):
@@ -144,5 +142,35 @@ class CategoryView(APIView):
         all_categories = Category.objects.filter(owner=request.user)
         serializer = CategorySerializer(all_categories, many=True, context={'request': request}) 
 
-        if all_categories:
+        if not all_categories.exists():
+            return Response({"status": "empty", "message": "No categories found."}, status=status.HTTP_204_NO_CONTENT)
+        else:
             return Response({"status": "success", "data": serializer.data}, status=status.HTTP_200_OK)
+
+
+class NoteTagSearchView(APIView):
+    authentication_classes = [JWTAuthentication]
+    permission_classes = [IsAuthenticated]
+    serializer_class = NoteSerializer
+
+    def get(self, request, tag):
+        try:
+            note = Note.objects.filter(tags__contains=tag, owner=request.user)
+            serializer = NoteSerializer(note, many=True, context={'request': request})
+
+            return Response({"status": "success", "data": serializer.data}, status=status.HTTP_200_OK)
+        
+        except Note.DoesNotExist:
+            return Response({"status": "error", "message": f"Notes with {tag} tag not found"}, status=status.HTTP_404_NOT_FOUND)
+        
+
+    def delete(self, request, tag):
+        try:
+            note = Note.objects.filter(tags__contains=[tag], owner=request.user)
+            if note.exists():
+                note.delete()
+                return Response({"status": "success", 'msg': 'Note deleted successfully'}, status=status.HTTP_204_NO_CONTENT)
+
+        except Note.DoesNotExist:
+            return Response({"status": "error", "message": f"Notes with {tag} tag not found"}, status=status.HTTP_404_NOT_FOUND)
+
